Remove code which can delete data in tables which is used in the create.database workflow. This is a functional change to create.database. But as the data loading using this code is ran just after creating a new database in practice it does not change any behavior.
As that table deletion code is the only user of the DatabaseFilter part of the dataio+sink
remove those from there as well.
Just keep one public method which is used from other code (in dbsm + in org.openbravo.ezattributes module).
Remove all body of the method as the only user of those filters was only being called from create.database
and it has been deleted (see above).
Mark that method as deprecated so callers can be identified and cleaned up over time as well.