Related to issue 22176: Detect removal of not null restrictions
The API check should detect the removal of a column not null restriction. If a column was not null someone could have taken for granted that that column was always going to have a not null value. After removing the not null restriction that column can now store null values, that could break the code that assumed that the column was going to store not null values.

--HG--
extra : rebase_source : e47fdf3939e5cce90ce9484f000a52fd4e5ff711