[func-idx] Code review changes (small refactor and log changes)
The isValidExpression was duplicated in PostgreSqlModelLoader and OracleModelLoader, so it has been moved to their parent class ModelLoaderBase. Also the isMonadicFunction method has been added to ModelLoaderBase, and now it is overriden in PostgreSqlModelLoader and OracleModelLoader.

Also the log has been improved (now Javadoc is used, and there are no e.printStackTrace();