[func-idx] Support more generic functions, not only those with one input param
The code that handles function based indexes has been modified so that is more flexible to accept more functions, not only monadic ones, for instance:

  CREATE INDEX c_bpartner_replaced_name
  ON c_bpartner
  USING btree
  (REPLACE(name, 'a', 'b'));

Now it is also supported to define indexes whose expression contains nested functions, for instance:

  CREATE INDEX c_bpartner_replaced_upper_name
  ON c_bpartner
  USING btree
  (UPPER(REPLACE(name, 'a', 'b')));

Now, instead of storing the name of the function applied to the index column, the whole expression is stored. The name of the index columns that use function expressions has been set to 'indexBasedColumn'.
Before storing the function expression it is transformed so that it is the same in Postgres and Oracle. In Postgres the cast expressions ('::text') are removed, along with the whitespaces, and also all the
 unquoted text in the expression is capitalized. In Oracle the quotes that surround the column name are removed.