[func-idx] Index column functions containing empty strings are not recreated
Oracle treats empty string as nulls. This behaviour resulted in the recreation of all the indexes whose columns used functions containing empty strings, because the comparison of the two functions (the one from the XML, with an empty strign, and the one from the database with null) resulted in a recreation.

This changeset fixes this, and now those indexes are not recreated. There is one pending issue: if the function used in an index contains an empty string as a parameter, when that index is exported there will be changes, as the empty string will be replaced by null in the XML. This will be addressed in a separate issue.