Fixes issue 31685: Wildcards are supported in exclude filter
Now it is supported to use wildcards [1] inside exclude filter, for instance:

<excludedTable name="SYM\_%"/>

That sentence will exclude all the tables whose smae starts with SYM_. It is supported for all the database objects that can be excluded: tables, views, triggers, functions and sequences.

An exclusion is considered to be a wildcard if it contains the '%' character, if not what will be excluded will be the database object whose name is equals to the provided name, ignoring the casing.

To escape the underscore character a backslash must be used, like in the example. The is logic inside ExcludeFilter and OracleModelLoader to handle how this must be escaped in Oracle (nothing else is required for Postgres).

[1] http://www.w3schools.com/sql/sql_wildcards.asp