[extend-exportsampledata] Dataset table where clause handling can be extended
In the ExportSampledata class the where clause of the dataset tables is replaced with a client filter. This is now done in a protected method, so classes that extend ExportSampledata can now manage the dataset table where clase in a different way (i.e. keeping the where clause defined in the AD dataset table).