[extend-exportsampledata] A table can be used several times in the same dataset
A subclass of ExportSampledata may need to support having several dataset tables that reference the same table (the most common case is when the excluded columns vary depending on the row being exported). This can't be done without adapting ExportSampledata to be extended.

The main behavioral change is that in the new task is a table is referenced in several dataset tables, the result of exporting those dataset tables will end up written in the same XML file.

To know whether a dataset table should be exported in a new XML file or in the previous one, the protected shouldExportToNewXMLFile(String tableName, String previousTableName) method has been defined. It returns true in ExportSampledata (this task assumes each table is referenced at most in one dataset table).