[extend-exportsampledata] Each dataset table is exported to a single XML file
This changeset reverts [1] and adds a new approach to support defining several times the same table in a dataset.

Before the current changeset all dataset tables referencing the same table would exports its contents to the same XML file. This required the structure of the task to change a lot. A much simpler approach is to export each dataset table to its own xml file, so if a table is used twice in a dataset, two files will be used to export it (named TABLE_NAME.xml, TABLE_NAME_1.xml, etc).

[1] https://code.openbravo.com/erp/devel/dbsm-extend-import-sample-data/rev/b2a9c4d0a3db