Related with issue 32553: Adds test that reenables the constraints of a table
The tests that where included with the issue [1] did not catch the current easy because they just disabled the constraints and foreign keys of tables, but they did not try to reenable them.

This changeset adds test to reenable the constraints and foreign keys of tables after having disabled them.

[1] https://issues.openbravo.com/view.php?id=31712