Fixed issue 32707: ORA-00918 error was get while packaging module
When a module was being packed in oracle, that error was shown.
This means that a column in ambiguously created.

The problem was in a query.
It was trying to execute a query where there were two tables with
a column with the same name. TABLE_NAME was being exectued insetad of
U.TABLE_NAME (with the alias of the table), and as that column exists in the two
tables, that error was raised.