related to issue 33964: Default value not dropped if column definition changes
The problem was caused because when the default value was a function (NOW()) this value was being parsed returning a null value. This was making impossible to detect the change in the default value by DB Source Manager. Now if the default value is function the value is not parsed.

Together with this, now when the default value is changed to NULL for a character varying column in Postgres the NULL value is translated properly within the PostgreSqlModelLoader.