Avoid problem with array and prepared statements by retrieving data properly
When using a prepared statement that returns a text array type in Postgres the prepared statement is reused starting from the fifth execution of the statement. From this point, the toString() method returns null for the Array as it is not implemented by the jdbc. To overcome this problem we use array.getArray() to retrieve the data and after that we print the comma separated list of values.