[prevent-cascade-constraint] The SqlBuilder.deleteInvalidConstraintRows changed
The SqlBuilder.deleteInvalidConstraintRows has been modified.
Before it was getting every foreign key of the table and after that was trying to delete all the records of the table which where referencing those foreign keys.
Now, it gets only the foreign keys of the tables which actually have a deletion. So, now the records deleted are those which are referencing those foreign keys, which belongs to tables which have had a deletion.