[partial-index] Updated method used to replace PostgresSQL cast expressions
The set of expressions that can include a cast in PostresSQL increases with the ability of defining partial indexes. For example the following expression: COL1 = 'Y' is casted in PostresSQL as follows: COL1::text = 'Y'::text. The replacement of those casts was not being covered by the algorithm implemented in removeCastExpressions() method of PostgreSqlModelLoader.

Now this method uses a more complete approach, based on regular expressions to find the cast expressions and replace them properly.