Related issue 33659: DB consistency error with function index (Oracle only)
When exporting database with a new custom function a db consistency error is
thrown in Oracle. The problem is that oracle add db user name at the beginning
of the indexExpression. For example, if "OBEQUALS(COL1)" is defined, export
database found this change "DB_USER.OBEQUALS(COL1)" and a db consistency error
is thrown. This problem was reproducible only in Oracle.

This problem has been solved removing db user name in the indexExpression in
OracleModelLoader.

Besides It has been created a test case for testing this
scenario in oracle and postgres too.