[prevent-cascade-constraint] Now the AD to AD tables cases are taken into account.
Before this fix, the AD to AD constrainst were always recreated and deleted, now some logic has been added. The constraints are only recreated when there in an insertion or a record deletion and not always. And the deletion of the FKs is only done when the table which is referenced by a FK has a deletion change.