[partial-index] Prevent losing oncreatedefault comment under some circumstances
In Oracle, when modifying the partial index information, if the column affected by that index already have an oncreatedefault comment, it was being deleted. This is now solved by keeping in the _onCreateDefaultColumns map all the changes in oncreatedefault comments during an update. This way it is easier to identify when the oncreatedefault has not been modified and in that case the related comment should never be dropped.