[dbsm-import-export-copy] Moves query gen logic to DataSetTableQueryGenerator
The logic to generate the queries to retrieve data from a dataset tables has been extracted to a new class called DataSetTableQueryGenerator. This allows to have all that logic centralized, and also allows to extend that class if a process that extends ExportSampledata needs to generate the queries in a different way.

The DatabaseDataIO and PgCopyDatabaseDataIO classes now accept a DataSetTableQueryGenerator object in their constructor, it instantiates the default one on the default constructor.

The ExportSampledata class now has a protected method called getQueryGenerator that allows the classes that extend it to define a custom query generator.