[dbsm-import-export-copy] All dsTables should have a table name
All the dataset tables should have its name property set to the name of the table they reference in order to be used in the query generator. Sometimes dataset table were passed to the DatabaseDataIO.readRowsFromTableList method without a table name. To prevent passing a dataset table without a name to the query generator, set in this point the dataset table name based on the table parameter that is also passed to the method.