[dbsm-import-export-copy] null moduleId cannot be replaced with an empty string
The DataSetTableQueryGenerator class was replacing null moduleId with an empty string before passing it to the dataSetTable.getWhereclause method. This was originally done because it was assumed it was the same.

But it is not the same. When the moduleId is passed as null to DatabaseDataIO.readRowsFromTableList (i.e. [1]) then it expects the dataset table where clause to be replaced with 1=1, but sometimes an empty string is passed ([2]) and it expects the dataset table where clause to be left as is. See the implementation of DatasetTable.getWhereClause: [3].

[1] https://code.openbravo.com/erp/devel/dbsm-main/file/1f2c47f076e2/src/org/openbravo/ddlutils/task/AlterDatabaseDataAll.java#l203
[2] https://code.openbravo.com/erp/devel/dbsm-main/file/1f2c47f076e2/src/org/openbravo/ddlutils/task/ExportSampledata.java#l105
[3] https://code.openbravo.com/erp/devel/dbsm-main/file/1f2c47f076e2/src/org/openbravo/ddlutils/util/OBDatasetTable.java#l46