[dbsm-import-export-copy] Change in whereClause handling in generated queries
Before this change, an additional where clause expression could be set in DataSetTableQueryGeneratorExtraProperties as a String, and it would be appended to the generated query for a dataset table. Instead of using a String, now this is defined as a List of columnName, operator and value. This will allow to support more complex queries, for instance using table aliases.