Related with issue 33875: Inlines method to avoid working with closed resultset
The DatabaseDataIO.getDatasetTableDataIterator method, which was only used in one place, returned a Iterator<DynaBean>, which was an iterator of a ResultSet. The problem was that the ResultSet was being closed inside the getDatasetTableDataIterator method, so the returned iterator could not work.

--HG--
extra : rebase_source : 130df4eaecfc04647b3d088dab15f02d42e6ea50