fixes issue 33997: DBSM not removing double precision cast properly
DBSM was not removing the double precision cast when exporting PostgreSQL model elements like views. The PostgreSQLStandarization is the class in charge of standarizing the PostgreSQL code which contains castings. But until now it was not considering the double precision numeric type.

Together with the fix a new suite intended for testing the views has been created as part of the DBSM test suite.