fixes bug 35370: Partial Index info lost under some circumnstances using Oracle
Partial Index information was not being saved properly in Oracle, after executing update.database having the following changes in the model:

1) A new column
2) A new partial index for the new column
3) Changes in another column of the same table which forces recreation

In this scenario, the partial index information was being overriden after executing the writeColumnCommentStmt method. To fix this problem now we keep the existing column comments at that point and concatenate them with the comments to be added to the new column (if any).

Note that this fix does not affect when there is no recreation because in this case, the partial index comments will be handled with the AddIndexChange model changes.