fixes issue 35634: Index generation in parallel fails under some circumstances
This problem can be reproduced when the number of max active connections of the pool used in DBSM is lower than the number of database connections retrieved in parallel to complete a task (currently the only task that performs queries in parallel is the creation of indexes on existing tables).
In order to prevent this problem, now the update.database and export.database tasks increase the number of active connections supported by the pool with a safe value: having one thread and the default max active connections (8) it was working fine, therefore we now set the max active connections for those tasks to the number of threads multiplied by 8.

--HG--
extra : rebase_source : 81bb760534ae1d5d9b4e27c158993956fe0a8b3c