Related with issue 35832: ImportSampleData now supports multithreading
The ImportSampleData class now supports defining the number of threads that will be used to import the files. This can be done by setting the max.threads parameter in the ant task that ends up invoking the ImportSampleData task, for instance:

ant install.source -Dmax.threads=1

It uses the default Platform's implementation, so if the max.threads parameter is not provided, it will use nAvailableCPUs/2.