Related with issue 35832: ExportSampleData now supports multithreading
The ExportSampleData class now supports defining the number of threads that will be used to import the files. This can be done by setting the max.threads parameter in the ant task that ends up invoking the ImportSampleData task, for instance:

ant export.sample.data -Dclient='The White Valley Group' -Dmodule=org.openbravo

It uses the default Platform's implementation, so if the max.threads parameter is not provided, it will use nAvailableCPUs/2.

Two additional changes have been done to support multithreading properly:
- To prevent mixing the log between threads, now the log does not show
   * Exporting table <tableName>
   * <n> records exported
   Now it will show <n> records exported from table <tableName>
- If the dataset being exported has more than one dataset table that references the same table, the first dataset table will be exported to <tableName>.xml and the second to <tableName>_1.xml. If no records are exported from the first dataset table <tableName>.xml will be deleted but the second one will still be named <tableName>_1.xml. With the previous version, given that the dataset tables were exported sequentially one at a time, it was checked if the previous one had  been deleted, and if so the second one would have been named <tableName>.xml. Now that several tables can be exported in parallel, each thread must know in advance the final name of the file where it will export data to.