related to bug 36137:ConfigScript with removeCheckChange on installsource fails
When applying data changes (RemoveCheckChange) of a Configuration Script on install.source
the changes was not take into account and no errors thrown.

This happend because code for install.source task is different than update.database task.
A design defect issue #34102 has been reported in order to fix this problem in a complete way.

The problem has been solved using the same logic which is being used on update.database. Now
applyConfigScriptModelChanges method is take into account RemoveCheckChange changes.