Related with issue 36678: Includes tableName parameter in transformWhereClause
Classes that extend DataSetTableQueryGenerator sometimes need to know the name of the table being exported in order to know how to transform its query. For instance, depending on the table being exported sometimes a clause to exclude the records defined for client 0.

Now that table name is passed as a parameter to transformWhereClause. It is an API change, but it does not have any consequences, I checked that no one else is invoking that method.

--HG--
extra : rebase_source : 2ee2a442eb0de48cc710ff7d157c7451299b5636