Related with issue 36908: Adds test
The test does the following:

A table T with a column with size 20 is created, along with three views with the following dependences:
- V1 -> T
- V2 -> V1
- V3 -> V2

Then the size of the column is updated from 20 to 40. Before issue 36908 was fixed the column size was not updated, and now it works properly.

Also, DbsmTest has been updated to ensure that continueOnError is true when alterTablesPostScript is invoked. This is needed because at that point the views are
recreated, and that step might need retries if some views depend on others.

--HG--
extra : rebase_source : 8d40bdc58f370db92a3181d3178a4fa5553c5c6b