Fixed issue 36950: CheckAPIDistribution task is not working as expected
The problem was introduced in these improvements of DBSM sources (#36137). The problem
was that CheckAPIDistribution task are not applied configScript files as expected and
for this reason, some warnings dissappear from the log in ret-api. Now the problem has
been fixed by applying the configScript files properly.

Besides it are reviewed the following tasks that used loadDataStructures method:
- ExportDatabase
- ExportConfigScript