Fixes issue 37103: import.sample.data fails if a file cannot be imported
If a file cannot be imported the import.sample.data process will stop immediately and will show log describing the problem.

To be able to catch the exceptions thrown in another thread, now the logic to import files is wrapped in a Callable instead of a Runnable.

--HG--
extra : rebase_source : b3cb0773c8eb407dabce404ed6037ad5c71745b1