[otherdbmodel] ExportDatabase can export model from other databases
ExportDatabase has been adapted so that now it can read the application dictionary data from an
(Openbravo) database, and the model from another database. This is useful for the reporting proyect,
since it enables defining the reporting model in an external (not Openbravo) database, and then
export it to a custom folder (i.e. src-db/database/reporting-model), while reading the Application
Dictionary data from an Openbravo database.

The following changes have been done:
- It no longer assumes it is being executed from an ant file defined in src-db/database, now it has
a proper way to find out where the Openbravo root path is
- Instead of having only one db (that contains both model and AD info), now it can have two. ExportDatabase
still only loads once the database, and reuse the same for both model and AD, but extensions can load
data from two different databases.
- The folder where model is exported (previously fixed to /src-db/database/model/) can now be defined
by extensions
- Adds a parameter to specify if AD data should be exported (some extensions might need to export
only the model (default to true). This parameter is also used to determine if the database validations
should include AD validations.
- Adds another parameter to specify if checksum info should be recalculated (default to true)